Bash completion with argcomplete
So, here is what argcomplete looks like. It is relatively easy but it
requires an additional "argcomplete" package available on pip.

Also, have you thought of adding an -m and --method option for HTTP
methods? Any specific reason it was not a part of option and just a
subcommand like syntax?