Make auth arg NOT req'd when using auth plugin
There are auth plugins that may want to load auth credentials
from the filesystem or environment variables.

This is a reasonable option for HMAC-based authentication schemes.
(Case in point: AWS APIs and SDKs work this way.)

So it enables a case where you specify an explicit --auth-type
but no --auth argument, passing None/None as username/password,
right to the auth plugin. It is then up to the auth plugin
to decide whether to load creds from somewhere, or instead
to throw a ValueError etc.

Tested against my fork of guardian/httpie-hmac-auth (forked
just to demonstrate):
    https://github.com/hangtwenty/httpie-hmac-auth/commit/2dee430a45c22aea49676e3e762744dcbcec47f7

More info incoming in PR.

Addresses #431.
Relates somewhat to:
- #378 at https://github.com/jkbrzt/httpie/issues/378
- https://github.com/teracyhq/httpie-jwt-auth/issues/3