Load cookies from response.headers instead of response.raw._original_response.msg._headers
response.cookies was not utilized as it not possible to construct original
payload from http.cookiejar.Cookie. Data is stored in lossy format. For example
Cookie.secure defaults to False so we cannot distinguish if Cookie.secure was
set to False or was not set at all. Same problem applies to other fields also.